refactor(streamwall): split main/data.ts into polling, watching, parsing and combining modules
main/data.ts sat in the top 3% by change entropy because it carried four
independent responsibilities at once, so unrelated changes kept landing in
the same file. Split it into a data/ module directory along those lines:

- data/poll.ts: pollDataURL with its fetch-timeout helpers
- data/watchFile.ts: watchDataFile
- data/parse.ts: payload validation (parseStreamEntries /
  parseStreamFileContents, extracted from the two identical inline blocks)
  and StreamIDGenerator
- data/combine.ts: markDataSource, latestPerSource, combineDataSources
- data/local.ts: LocalStreamData and presetDataSource
- data/types.ts: the shared DataSource/health-callback types

data/index.ts re-exports the exact previous public surface, so every
'./data' import (including the vi.mock in index.test.ts) is untouched --
the same directory-with-index pattern main/presets already uses. The test
file is split alongside the modules; all 48 tests moved verbatim, plus two
new ones for the extracted parseStreamEntries. No behavior change.
